+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - Plugin module name convention is now enforced.
 - The `logs` directory is now always created relative to the corpus directory instead of the directory where `sparv` is
   run.
+- Detailed error messages are now always printed to the log file, no matter the log level.
 
 ### Fixed